requirements.txt
ultralytics
torch
opencv-python
numpy
matplotlib
pillow